+++ packages/contracts/test/global_hooks.ts
@@ -1,7 +1,7 @@
 import { getCoverageSubproviderSingleton } from '@0xproject/dev-utils';
 
 after('generate coverage report', async () => {
-    if (process.env.COVERAGE) {
+    if (process.env.SOLIDITY_COVERAGE) {
         const coverageSubprovider = getCoverageSubproviderSingleton();
         await coverageSubprovider.writeCoverageAsync();
     }