+++ packages/0x.js/monorepo_scripts/globals.d.ts
@@ -1,9 +1,12 @@
-declare module 'web3-provider-engine';
-declare module 'web3-provider-engine/subproviders/rpc';
 declare module 'async-child-process';
 declare module 'publish-release';
 
 // semver-sort declarations
 declare module 'semver-sort' {
     const desc: (versions: string[]) => string[];
+}
+
+declare module '*.json' {
+    const value: any;
+    export default value;
 }