+++ packages/dev-utils/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v0.3.0 - _TBD, 2018_
 
-    * Add coverage subprovider if enabled (#426)
+    * Add coverage subprovider if SOLIDITY_COVERAGE env variable is true (#426)
     * Refactor `BlockchainLifecycle` to work with in-process ganache (#426)
     * Remove `RPC` class and move it's logic to `Web3Wrapper` (#426)